simplified a couple things; updated to work again with Qt 6.2
emissiveFactor is now a vector3d being used as the RGB components,
aka a color.  Seems weird but was done to match glTF2's concept:
https://codereview.qt-project.org/c/qt/qtquick3d/+/334411
And it makes self-illuminating 2D subscenes some hassle now.

If diffuseMap has a 2D scene, event delivery is fine, but emissive color
overwhelms it rather than just making non-black colors brighter.
So it seems we need to do something with emissiveMap.  Setting
emissiveMap: diffuseMap
seems to make it self-illuminate about like it was before.
But if the 2D scene is only set on emissiveMap, event delivery doesn't work.
(And that's apparently intentional too.)

Other trickery is possible, with custom fragment shaders and such.